automations/requirements.txt
flask
APScheduler
Flask-APScheduler
python-dotenv
pandas
google
google-auth
google-auth-oauthlib
google-api-python-client
pymongo